requirements.txt
adafruit-circuitpython-pn532
usbrelay-py
RPi.GPIO
rpi_ws281x
neopixel
smbus2
requests
RPLCD
PyYAML 